requirements.txt
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
Pillow==10.1.0
psycopg2-binary==2.9.9
pytest-django==4.7.0
python-dotenv==1.0.1
gunicorn==21.2.0
